% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{as_glaredb_table}
\alias{as_glaredb_table}
\title{Create a GlareDB table}
\usage{
as_glaredb_table(x, ...)
}
\arguments{
\item{x}{An object to be coerced to a GlareDB table.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A GlareDB table.
}
\description{
Create a GlareDB table
}
\examples{
con <- glaredb_connect()
dat <- as_glaredb_table(data.frame(a = 1:3, b = letters[1:3]))

glaredb_sql("SELECT * FROM dat", con) |>
  as.data.frame()
}
