% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{as_glaredb_table}
\alias{as_glaredb_table}
\alias{RGlareDbTable}
\alias{as_glaredb_table.default}
\alias{as_glaredb_table.nanoarrow_array_stream}
\alias{as_glaredb_table.RGlareDbExecutionOutput}
\title{Create a GlareDB table}
\usage{
as_glaredb_table(x, ...)

\method{as_glaredb_table}{default}(x, ...)

\method{as_glaredb_table}{nanoarrow_array_stream}(x, ...)

\method{as_glaredb_table}{RGlareDbExecutionOutput}(x, ...)
}
\arguments{
\item{x}{An object to be coerced to a GlareDB table.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A GlareDB table.
}
\description{
A class that has a struct similar to \link[arrow:Table-class]{arrow::Table} innerly and
can be converted to other classes via \code{\link[nanoarrow:as_nanoarrow_array_stream]{nanoarrow::as_nanoarrow_array_stream()}}.
}
\examples{
con <- glaredb_connect()
dat <- as_glaredb_table(data.frame(a = 1:3, b = letters[1:3]))

glaredb_sql("SELECT * FROM dat", con) |>
  as_glaredb_table()
}
