% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glaredb-package.R
\docType{package}
\name{glaredb-package}
\alias{glaredb}
\alias{glaredb-package}
\title{glaredb: R Bindings for 'GlareDB'}
\description{
TODO
}
\author{
\strong{Maintainer}: Tatsuya Shima \email{ts1s1andn@gmail.com}

Authors:
\itemize{
  \item Authors of the dependency Rust crates (see inst/AUTHORS file for details)
}

}
\keyword{internal}
