% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glaredb-package.R
\docType{package}
\name{glaredb-package}
\alias{glaredb}
\alias{glaredb-package}
\title{glaredb: R Bindings for 'GlareDB'}
\description{
R bindings for 'GlareDB', an analytical database based on 'Apache Arrow' and 'Apache DataFusion' that can connect to various data sources and execute queries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://eitsupi.github.io/r-glaredb/}
  \item \url{https://github.com/eitsupi/r-glaredb}
  \item Report bugs at \url{https://github.com/eitsupi/r-glaredb/issues}
}

}
\author{
\strong{Maintainer}: Tatsuya Shima \email{ts1s1andn@gmail.com}

}
\keyword{internal}
