% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{glaredb_connect}
\alias{glaredb_connect}
\alias{RGlareDbConnection}
\title{Connect to a GlareDB database}
\usage{
glaredb_connect(
  data_dir_or_cloud_url = NULL,
  ...,
  spill_path = NULL,
  disable_tls = FALSE,
  cloud_addr = "https://console.glaredb.com",
  location = NULL,
  storage_options = NULL,
  env = parent.frame()
)
}
\arguments{
\item{data_dir_or_cloud_url}{A character of path to a local GlareDB database or
a cloud URL or \code{NULL}. If \code{NULL}, a in-memory database is used.}

\item{...}{Ignored.}

\item{spill_path}{TODO}

\item{disable_tls}{\code{TRUE} or \code{FALSE} to indicating whether to disable TLS.}

\item{cloud_addr}{A character of a GlareDB cloud URL.}

\item{location}{TODO}

\item{storage_options}{Named character vector of storage options or \code{NULL} (default).}

\item{env}{TODO}
}
\value{
GlareDB connection object
}
\description{
Connect to a GlareDB database
}
\examples{
con <- glaredb_connect()
con

glaredb_sql("SELECT 'hello from R' as hello", con) |>
  as_glaredb_table()
}
