% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R, R/execution.R
\name{glaredb_sql}
\alias{glaredb_sql}
\alias{glaredb_prql}
\alias{glaredb_execute}
\title{Run a query against a GlareDB database}
\usage{
glaredb_sql(query, connection = NULL)

glaredb_prql(query, connection = NULL)

glaredb_execute(query, connection = NULL)
}
\arguments{
\item{query}{A character of the query to run.
\itemize{
\item For \code{\link[=glaredb_sql]{glaredb_sql()}} and \code{\link[=glaredb_execute]{glaredb_execute()}}, an SQL query.
\item For \code{\link[=glaredb_prql]{glaredb_prql()}}, a PRQL query.
}}

\item{connection}{A \link[=glaredb_connect]{GlareDB connection object} or \code{NULL}.
If \code{NULL}, the default in-memory database is used.}
}
\value{
A GlareDB execusion output (Query plan).
For \code{\link[=glaredb_execute]{glaredb_execute()}}, the value is returned invisibly.
}
\description{
Run a query against a GlareDB database
}
\examples{
# You can materialize the query result by `as_glaredb_table()` etc.
glaredb_sql("SELECT 'hello from R' as hello") |>
  as_glaredb_table()

glaredb_prql("from [
  {a=5, b=false},
  {a=6, b=true},
]") |>
  as.data.frame()

# `glaredb_execute()` is useful for manipulating the database
glaredb_execute("CREATE TABLE my_table (a int)")
glaredb_execute("INSERT INTO my_table VALUES (1), (2)")

glaredb_sql("SELECT * FROM my_table") |>
  as_glaredb_table()
}
