% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{glaredb_sql}
\alias{glaredb_sql}
\title{Run a SQL query against the default in-memory GlareDB database}
\usage{
glaredb_sql(query)
}
\arguments{
\item{query}{A character of the SQL query to run}
}
\value{
GlareDB execusion output
}
\description{
Run a SQL query against the default in-memory GlareDB database
}
\examples{
glaredb_sql("SELECT 'hello from R' as hello") |>
  as.data.frame()
}
